rename image files to fanart
